Updated data migration format
This is an update to the format of data specified in the schema.json file for migration.  The issue was doing updates to multiple rows in a single table and applying some ordering.  This is useful for resolving circular dependencies.

For example, an "account" table references a "user" table for specifying an admin user.  The "user" table then references the "account" table to specify the account that user record is linked to.  To solve this issue  you have to insert all the "account" records, Then insert all the "user" records.  Then update the "account" record to link to the "user" records.  This is now possible.