Added data update directive
The default action for data records is to synchronise them.  There are two optional parameters "insertonly" and "updateonly" that can be used to modify how that workds.  But we now have an "update" directive as well which allows multiple rows to be updated in one go based on matching criteria.