MAJOR: Reversed the assoc field specification
Previously in the fields array, the key would be the source column and the value would be the AS part in the SQL.

So array('user_id' => 'id') would result in: "SELECT user_id AS id".

When we started writing complex fields, such as array("EXTRACT(EPOCH FROM my_timestamp)::integer" => 'seconds') it made less sense to do it that way.  So I have simply flipped them around.

This could have quite an impact on existing code.  I know personally I don't use this feature very often so I'm ok.  haha.