Fixed up data sync on migration
Now sure why this wasn't comitting.  It was executing the code but was always doing a rollback.  I moved the transaction handling into the sync method itself and it's OK now.

Also noticed that moving the data schema into separate files would cause this to not run as well so fixed that up too.

fixes #8