Fixed ROOT object re-create filesystem check
Previously when we re-created the root object we just updated all existing nodes to point to it, which meant we lost the filesystem structure entirely.  Now instead we just run an fsck() which is a bit smarter and only updates missing parents while attempting to maintain good working structure.

fixes #19