Added master server support
This is a feature we had ages ago for PGSQL and was lost during the DBD rewrite.  I have re-implemented this but in the BaseDriver so that it can be used on all DBD drivers.  

To use this, specify a 'master' config option which contains the name of the master server.  Because this is designed to be used with existing server replication functionality, the only option required is the host name and all other details (username, password, driver options, etc) will be used from the base connection.

Currently only tested on PGSQL but it works as described.  So any INSERT, UPDATE or DELETE queries will go to the master.