Enhanced JSON column data sync
Diff was previously only detecting added elements because this is how array_diff_assoc_recursive() works.  It will only return elements that do not exist in the source array.  So detect things that have been removed, we now do an additional reverse check if there are already no changes detected (if there are already changes the update will always work).