Added a fancy Row class
This class is used to access a result row returned by a query.  It makes use of the statement column metadata to setup a strict model for strict typing the column values.  

This release also has a preliminary update() method that allows the row to be updated.  This includes support for JOINs and will trigger an update on each table that has changed data as part of the join.