Fixed migration not working for some reason
Turned out to be a change to a default value on the insert chain of methods.  Returning was default to NULL on the Adapter class, but is TRUE on the BaseDriver and they do slightly different things.  The problem was only because TRUE get the last insert primary key and the schema_info table has no sequence on the primary key.