fetch()/fetchall() legacy support and enhanced encryption
This is a major update to the new 2.0 development branch.  The major changes here are to the philosophy used to decide when a new Row object should be returned.  Now we only return these objects when using the row(), rows() or result array access methods.  Using fetch() and fetchall() continues to return an array.

This required major updates to the way the encrypted fields work.  The encryption is now in the adapter itself and can intercept any field when doing in insert or update.  The decryption is handled in the result because the Result object has access to the column metadata.