File Backend updated
The issue here is the database layout has changed slightly.  Instead of chunks linking to files, it is reversed.  This means I am going to have to come up with an upgrade path for existing storage.

The benefit now is that files are independent of each other so they can be renamed individually even though they point to the same data chunk.  They can also have different properties, meta, etc.  When the contents is updated it will be rechecked with MD5 signature and linked to a new chunk, or insert a new chunk and linked to that.