Added "keep tables" option
This  allows building a schema on an existing database.  It checks if any tables that are about to be created the tables exist and if so, does a diff on them to see if they are using the current schema.  If they are not, an error will be thrown and migration will stop.  If the table/constraint/view/function/trigger exists and is the same as the current schema version, then a note is made in the log and things carry on.