DBD BaseDriver now sets timezone
We are sending the config in the constructor now so that it is always kept in the driver, allowing driver options to be set more easily.  Because of this, the driver now sets the timezone so that we don't have to check for a master and set it there too.