Added support for subquery fields
You can now pass a DBI query ie: \Hazaar\DBI\Table, as a field and it will be automatically output as a string and aliased.  If the subquery has a limit of 1, then the field will be the returned value, otherwise the field will be an array of all rows.  As per PGSQL only one column can be selected however we do not enforce this so that the standard SQL error is shown.