Data sync macros
First pass at this is now working.  We are using SQL queries to lookup though, which I think is inefficient.  I think it would be cool to use stored row data if it is available, with the SQL lookup used as a fallback if no data exists.  This will add efficiency by not having to do queries for things that have just been inserted/updated, but also allow non-synced data to be used in macros.