Stage 1: Build schema from migration files
I seem to have this working.  This function replaces the code that loads the schema file and instead loads it by replaying the migration files.  Will take more time, but should not be much and the benefits of this will be huge.  

Full support for tables, constraints and views ONLY at this point.

Still need to add support for indexes and functions.