Removed migration "version" attribute and Major fixes
This is more of a major fix for the new schema generator than anything.  It was quite simple to remove the version attribute.  Changing the JSON parser to use objects would not have worked, so I am using the `$table_map` variable to get a list of valid level1 attributes (tables, indexes, etc) and using that to determine which way to call the replay.

Also, massive fixes to the alter stuff in the schema generator.  Turns out if you alter some things in tables, you need to automatically alter them in other places, like constraints and indexes.  We are now doing this.

I have tested this against a clients migration files which are quite complex and it was able to initialise a database at various versions as well as replay every migration.  Can't get much more betterer than that.