Added support for RETURNING multiple columns on INSERT
Previous functionality remains consistent.  So if the $returning parameter is NULL or FALSE, the number of affected rows is returned.  If it is TRUE then db->lastInsertId() is returned.  If it is a string naming a column then that column value will be returned.

New functionality allows this parameter to be an array.  In this case, an array is returned which is an array of column values from the columns named in the $returning parameter.