Fixed creating directory that was created after directory load
Occasionally, concurrent requests may result in a directory being created at the same time.  The first one to do so will check if it exists then if not create it.  The second could potentially not see the existing directory because it get created in the time between when it checks if it exists, and when it goes to create it itself.  This throws an error, which we now look for specifically and reload the parent directory info so that subsequent function calls know the path now exists.