Fixed array/json columns from join which have same name
Doing a select group, with array columns from multiple tables where each table may contain a column that shares a name was not working.  The chances of this occurring are slim, but I have tested with two tables both sharing a JSON field named 'params'.  Previously, the record fixer would not drop into the next level so all JSON would be an array and the subsequent would be missing.  Now we can tell if there are actually multiple fields returned with the same name and decode them individually.