Added basic instance tracking to Hazaar\DBI\Adapter
This will track the last instance of an adapter.  It's safe to call getInstance() always as this will  allow reusing a single connection.  However, if a second connection is desired it is still possible to instantiate another instance of an Adapter directly, however these subsequent instances will not be tracked.